Ajout de getters dans city sur inhabitants
Ajout d'un d pour la valeur double bankAccount dans Inhabitant

Changement du type de amount dans PromissoryNote

Modification du main pour créer une simulation

Création d'une classe Simulation qui va simuler ce qui est demandé

+++ Projet 3 Courriers/src/model/content/letter/PromissoryNote.java
@@ -5,7 +5,7 @@ import model.content.Money;
 
 public class PromissoryNote extends Letter<Money> {
 	
-	public PromissoryNote(Inhabitant sender, Inhabitant receiver, int amount) {
+	public PromissoryNote(Inhabitant sender, Inhabitant receiver, double amount) {
 		super(sender, receiver, new Money(amount));
 	}
